# Conflicts:
#	server/src/main/java/de/ingrid/igeserver/api/UsersApiController.kt
#	server/src/main/java/de/ingrid/igeserver/repository/UserRepository.kt
#	server/src/main/java/de/ingrid/igeserver/services/CatalogService.kt

+++ frontend/src/app/+research/+query-manager/query-manager.component.ts
@@ -1,6 +1,9 @@
 import {Component, EventEmitter, OnInit, Output} from '@angular/core';
 import {QueryQuery} from '../../store/query/query.query';
 import {ResearchService} from '../research.service';
+import {ConfirmDialogComponent, ConfirmDialogData} from '../../dialogs/confirm/confirm-dialog.component';
+import {MatDialog} from '@angular/material/dialog';
+import {Query} from '../../store/query/query.model';
 
 @Component({
   selector: 'ige-query-manager',
@@ -12,11 +15,10 @@ export class QueryManagerComponent implements OnInit {
   @Output() selection = new EventEmitter<string>();
 
 
-  systemQueries = this.queryQuery.selectAll();
-
-  userQueries = [];
+  userQueries = this.queryQuery.selectAll();
 
   constructor(private queryQuery: QueryQuery,
+              private dialog: MatDialog,
               private researchService: ResearchService) {
   }
 
@@ -24,10 +26,27 @@ export class QueryManagerComponent implements OnInit {
   }
 
   removeQuery(id: string) {
-    this.researchService.removeQuery(id).subscribe();
+    this.dialog.open(ConfirmDialogComponent, {
+      data: <ConfirmDialogData>{
+        message: `Möchten Sie die Anfrage wirklich löschen`,
+        title: 'Löschen',
+        buttons: [
+          {text: 'Abbrechen'},
+          {text: 'Löschen', alignRight: true, id: 'confirm', emphasize: true}
+        ]
+      }
+    }).afterClosed().subscribe(result => {
+      if (!result) return;
+
+      this.researchService.removeQuery(id).subscribe();
+    });
   }
 
   load(id: string) {
     this.selection.emit(id);
+  }
+
+  getIdentifier(index, item: Query) {
+    return item.id;
   }
 }
